# analytic-service/src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: Tron Analytics API
  description: API for analyzing Tron blockchain contracts and transactions
  version: 1.0.0
  contact:
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api.example.com/api
    description: Production server
tags:
  - name: User
    description: Operations about users
  - name: Analytics
    description: Blockchain contract and transaction analytics
paths:
  /getUser:
    get:
      tags:
        - User
      summary: Get user by identifiers
      description: Returns a user by either their Telegram ID or Tron wallet address
      operationId: getUserId
      parameters:
        - name: telegramId
          in: query
          description: Telegram ID of the user
          required: false
          schema:
            type: string
            example: "123456789"
        - name: walletAddress
          in: query
          description: Tron wallet address of the user
          required: false
          schema:
            type: string
            example: "TXYZ123..."
      responses:
        '200':
          description: User found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request - neither parameter provided
          content:
            application/json:
              schema:
                type: string
                example: "At least one of telegramId or walletAddress must be provided"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/tronscan/{tokenAddress}:
    get:
      tags:
        - Analytics
      summary: Get token statistics
      description: Returns comprehensive statistics for a TRON token contract
      operationId: getTokenStats
      parameters:
        - name: tokenAddress
          in: path
          description: TRON token contract address
          required: true
          schema:
            type: string
            example: "TR7NHqjeKQxGTCi8q8ZY4pL8otSzgjLj6t"
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenStatsDTO'
        '400':
          description: Invalid token address
        '404':
          description: Token not found
        '500':
          description: Internal server error

  /api/v1/transaction/{transactionHash}:
    get:
      tags:
        - Analytics
      summary: Get transaction analysis
      description: Returns detailed analysis of a TRON transaction
      operationId: getTransactionAnalysis
      parameters:
        - name: transactionHash
          in: path
          description: TRON transaction hash
          required: true
          schema:
            type: string
            example: "0x123abc..."
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionAnalysisDTO'
        '400':
          description: Invalid transaction hash
        '404':
          description: Transaction not found
        '500':
          description: Internal server error

  /users/history:
    get:
      tags:
        - User
      summary: Get user history
      description: Returns the transaction history for a user identified by either telegramId or walletAddress, saves only past 10 queries
      operationId: getUserHistory
      parameters:
        - name: telegramId
          in: query
          description: Telegram ID of the user
          required: false
          schema:
            type: string
        - name: walletAddress
          in: query
          description: Wallet address of the user
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RequestDTO'
        '400':
          description: Invalid parameters
        '404':
          description: User not found
        '500':
          description: Internal server error

  /users:
    post:
      tags:
        - User
      summary: Create a new user
      description: Creates a new user with the provided details
      operationId: createUser
      parameters:
        - name: name
          in: query
          description: Name of the user
          required: true
          schema:
            type: string
        - name: telegramId
          in: query
          description: Telegram ID of the user
          required: false
          schema:
            type: string
        - name: walletAddress
          in: query
          description: Wallet address of the user
          required: false
          schema:
            type: string
      responses:
        '200':
          description: User created successfully
        '400':
          description: Invalid input
        '500':
          description: Internal server error

    put:
      tags:
        - User
      summary: Merge user data
      description: Merges user data identified by name with either telegramId or walletAddress
      operationId: mergeUser
      parameters:
        - name: name
          in: query
          description: Name of the user
          required: true
          schema:
            type: string
        - name: telegramId
          in: query
          description: Telegram ID of the user
          required: false
          schema:
            type: string
        - name: walletAddress
          in: query
          description: Wallet address of the user
          required: false
          schema:
            type: string
      responses:
        '200':
          description: User data merged successfully
        '400':
          description: Invalid input
        '500':
          description: Internal server error

    delete:
      tags:
        - User
      summary: Delete a user
      description: Deletes a user identified by either telegramId or walletAddress
      operationId: deleteUser
      parameters:
        - name: telegramId
          in: query
          description: Telegram ID of the user
          required: false
          schema:
            type: string
        - name: walletAddress
          in: query
          description: Wallet address of the user
          required: false
          schema:
            type: string
      responses:
        '200':
          description: User deleted successfully
        '400':
          description: Invalid parameters
        '500':
          description: Internal server error

  /users/{userId}/requests:
    post:
      tags:
        - User
      summary: Save user request
      description: Saves analysis results for a user's request, body parameters depends on what query user send. if User send get Get token statistics so the addressResult need to be fullfilled and transactionResult would be empty
      operationId: saveUserRequest
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Analysis results to save
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                addressResult:
                  $ref: '#/components/schemas/TokenStatsDTO'
                transactionResult:
                  $ref: '#/components/schemas/TransactionAnalysisDTO'
      responses:
        '200':
          description: Request saved successfully
        '400':
          description: Invalid input
        '404':
          description: User not found
        '500':
          description: Internal server error

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          description: Unique identifier of the user
        name:
          type: string
          example: "John Doe"
          description: Name of the user
          maxLength: 100
        telegramId:
          type: string
          example: "123456789"
          description: Unique Telegram ID of the user
          nullable: true
        tronWallet:
          type: string
          example: "TXYZ123..."
          description: Unique Tron wallet address
          nullable: true
        requests:
          type: array
          items:
            $ref: '#/components/schemas/RequestDTO'
          description: List of requests associated with the user
      required:
        - id
        - name

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2023-10-01T12:00:00Z"
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "NOT_FOUND"
        message:
          type: string
          example: "User not found with provided identifiers"
        path:
          type: string
          example: "/getUserId"
      required:
        - timestamp
        - status
        - error
        - message

    RequestDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
          description: Request ID
        hash:
          type: string
          example: "0x123abc"
          description: Transaction hash
        address:
          type: string
          example: "TXYZ123..."
          description: Wallet address
        addressResult:
          $ref: '#/components/schemas/TokenStatsDTO'
        transactionResult:
          $ref: '#/components/schemas/TransactionAnalysisDTO'
      required:
        - id
        - hash
        - address

    TokenStatsDTO:
      type: object
      properties:
        name:
          type: string
          example: "Tether USD"
          description: Token name
        tokenType:
          type: string
          example: "TRC20"
          description: Type of token (TRC10, TRC20, etc.)
        logo:
          type: string
          example: "https://example.com/logo.png"
          description: URL to token logo
        totalSupply:
          type: string
          example: "1000000000"
          description: Total token supply (as string to handle large numbers)
        projectPage:
          type: string
          example: "https://tether.to"
          description: Project homepage URL
        marketCap:
          type: string
          example: "80000000"
          description: Market capitalization in USD
        holders:
          type: string
          example: "50000"
          description: Number of token holders
        totalTransfers:
          type: string
          example: "1000000"
          description: Total number of transfers
        transfersYesterday:
          type: string
          example: "5000"
          description: Number of transfers in last 24 hours
        volume24h:
          type: string
          example: "1000000"
          description: Trading volume in last 24 hours
        liquidity:
          type: string
          example: "500000"
          description: Liquidity amount
        contractAddress:
          type: string
          example: "TR7NHqjeKQxGTCi8q8ZY4pL8otSzgjLj6t"
          description: Token contract address
        creationDate:
          type: string
          format: date-time
          example: "2021-01-01T00:00:00Z"
          description: Token creation date
      required:
        - name
        - tokenType
        - contractAddress

    # TokenStatsDTO:
    #   type: object
    #   properties:
    #     tokenBalance:
    #       type: number
    #       format: double
    #       example: 1000.5
    #       description: Token balance
    #     tokenValueUSD:
    #       type: number
    #       format: double
    #       example: 1500.75
    #       description: Token value in USD
    #     transactionsCount:
    #       type: integer
    #       example: 42
    #       description: Number of transactions
    #   required:
    #     - tokenBalance
    #     - tokenValueUSD
    #     - transactionsCount

    TransactionAnalysisDTO:
      type: object
      properties:
        transactionHash:
          type: string
          example: "0x123abc..."
          description: Hash of the transaction
        timestamp:
          type: string
          example: "2023-10-01T12:00:00Z"
          description: Timestamp of the transaction
        blockNumber:
          type: string
          example: "12345678"
          description: Block number containing the transaction
        analysis:
          $ref: '#/components/schemas/Analysis'
      required:
        - transactionHash
        - analysis

    Analysis:
      type: object
      properties:
        type:
          type: string
          example: "TRANSFER"
          description: Type of transaction analysis
        mainContract:
          type: string
          example: "0xcontract123..."
          description: Main contract address involved
        callHierarchy:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/CallNode'
          description: Hierarchical call structure
        eventLogs:
          type: array
          items:
            $ref: '#/components/schemas/EventLog'
          description: List of event logs
      required:
        - type
        - callHierarchy
        - eventLogs

    CallNode:
      type: object
      properties:
        hash:
          type: string
          example: "0xcallhash123..."
          description: Hash of the call
        caller:
          type: string
          example: "0xcaller123..."
          description: Caller address
        callee:
          type: string
          example: "0xcallee123..."
          description: Callee address
        note:
          type: string
          example: "Token transfer"
          description: Additional notes about the call
        value:
          type: string
          example: "1000000000000000000"
          description: Value transferred in the call (as string to handle large numbers)
        children:
          type: array
          items:
            $ref: '#/components/schemas/CallNode'
          description: Nested calls
      required:
        - hash
        - caller
        - callee

    EventLog:
      type: object
      properties:
        contract:
          type: string
          example: "0xcontract123..."
          description: Contract address that emitted the event
        eventType:
          type: string
          example: "Transfer"
          description: Type of the event
        decoded:
          type: boolean
          example: true
          description: Whether the event was successfully decoded
        signature:
          type: string
          example: "Transfer(address,address,uint256)"
          description: Event signature
        possibleEvent:
          type: string
          example: "Transfer"
          description: Possible event name if not fully decoded
        event:
          type: string
          example: "Transfer"
          description: Final determined event name
        from:
          type: string
          example: "0xsender123..."
          description: Sender address
        to:
          type: string
          example: "0xreceiver123..."
          description: Receiver address
        value:
          $ref: '#/components/schemas/Value'
        rawData:
          type: string
          example: "0xdata123..."
          description: Raw event data
        rawTopics:
          type: array
          items:
            type: string
          example: ["0xtopic1", "0xtopic2"]
          description: Raw event topics
        sender:
          type: string
          example: "0xsender123..."
          description: Event sender
        amount0In:
          type: object
          description: Amount in for token 0 (can be number or string)
          nullable: true
        amount1In:
          type: object
          description: Amount in for token 1 (can be number or string)
          nullable: true
        amount0Out:
          type: object
          description: Amount out for token 0 (can be number or string)
          nullable: true
        amount1Out:
          type: object
          description: Amount out for token 1 (can be number or string)
          nullable: true
      required:
        - contract
        - decoded

    Value:
      type: object
      properties:
        raw:
          type: string
          example: "1000000000000000000"
          description: Raw value as string
        formatted:
          type: string
          example: "1.0"
          description: Formatted human-readable value
        trx:
          type: string
          example: "1 TRX"
          description: Value with currency symbol
      required:
        - raw

    Amount:
      type: object
      properties:
        raw:
          type: number
          format: double
          example: 1.0
          description: Raw numeric value
        formatted:
          type: string
          example: "1.0"
          description: Formatted human-readable value
        trx:
          type: string
          example: "1 TRX"
          description: Value with currency symbol
      required:
        - raw
        - formatted

    Error:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2023-10-01T12:00:00Z"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Invalid input data"
        path:
          type: string
          example: "/api/users"